Added GetMember() to IFamily. Moved JToken and custom System/Family instantiation into JsonEngine. Renamed instance to singleton for Root and Engine. Updated Transform2D with global and local matrices. Overhaul of Render2DSystem to handle scenegraph.

+++ ECS/Objects/EngineUpdater.cs
@@ -6,10 +6,10 @@ namespace Atlas.ECS.Objects
 	public class EngineUpdater
 	{
 		private readonly Stopwatch timer = new Stopwatch();
-		private readonly Action<float> update;
+		private readonly Action<double> update;
 		private bool isRunning = false;
 
-		public EngineUpdater(Action<float> update)
+		public EngineUpdater(Action<double> update)
 		{
 			this.update = update ?? throw new NullReferenceException();
 		}
